requirements.txt
Cheetah==2.4.4
Flask==0.10.1
Flask-HTTPAuth==3.2.1
Flask-Login==0.4.0
Flask-SQLAlchemy==2.1
Flask-WTF==0.13.1
Jinja2==2.7.2
Landscape-Client==14.12
MarkupSafe==0.18
PAM==0.4.2
PyYAML==3.10
SQLAlchemy==1.1.4
SQLAlchemy-Utils==0.32.11
SecretStorage==2.0.0
Twisted-Core==13.2.0
Twisted-Names==13.2.0
Twisted-Web==13.2.0
WTForms==2.1
WTForms-Alchemy==0.16.1
WTForms-Components==0.10.1
Werkzeug==0.9.4
apt-xapian-index==0.45
argparse==1.2.1
bleach==1.5.0
blinker==1.3
chardet==2.0.1
click==6.6
cloud-init==0.7.5
colorama==0.2.5
configobj==4.7.2
decorator==4.0.10
html5lib==0.999
httplib2==0.9.2
infinity==1.4
intervals==0.8.0
itsdangerous==0.22
jsonpatch==1.3
jsonpointer==1.0
keyring==3.5
launchpadlib==1.10.2
lazr.restfulclient==0.13.3
lazr.uri==1.0.3
oauth==1.0.1
oauth2client==4.0.0
passlib==1.6.5
prettytable==0.7.2
psycopg2==2.4.5
pyOpenSSL==0.13
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycrypto==2.6.1
pycurl==7.19.3
pygobject==3.12.0
pyinotify==0.9.4
pyserial==2.6
python-apt==0.9.3.5ubuntu2
python-debian==0.1.21-nmu2ubuntu2
python-dotenv==0.6.1
rauth==0.7.2
redis==2.10.5
requests==2.2.1
rsa==3.4.2
simplejson==3.3.1
six==1.10.0
ssh-import-id==3.21
urllib3==1.7.1
validators==0.11.1
wadllib==1.3.2
wheel==0.24.0
wsgiref==0.1.2
zope.interface==4.0.5
